Sharded directors support (#122)
* Pass an another variable to go templates. It's called `Frontends` and has the information about varnish pods deployed. Needed for shard director which needs to know about all varnish nodes in the cluster.
Deleted kwatcher/backends package as it's function were not used anywhere. Moved the struct to the controller package.
Moved label names constants to the apis/icm/v1alpha1 package. That way we can reuse those constants in kwatcher and varnishservice. Also its part of the API so it belongs there anyway.
Remove HostPort usage in deployment. It make kubernetes open that port on every node. That leads to issue when you can't create more than one pod as the first pod opens the port and the rest of them hang in `Pending` state as no ports available on the nodes.

* Fix endpoints watching.
Give every component name a constant.

* Add yaml manifests for installing promeheus operator, prometheus itself for varnish monitoring and grafana.
Add config map for sharded Varnish.

* Move labelmaker to a separate package and use it in kwatcher and operator.
Prevent pods trigger reconcile loops and varnish reloads for each other. That led to potentially infinate VCL reload loop.
Fixed by not taking into consideration readiness of an endpoint. Also the list of endpoints are sorted to prevent generation of different VCL files for the same set of endpoints that differ only by order in the slice.

* Fix the less function and make it more concise.

* Log if the object type is wrong in the predicate.
Inherit varnish service labels for pods.

* Bump version.