Improve logging. (#144)
* Use context for passing the logger. That allows to have inherited logger where you can have already set fields in your logger and complement it with additional fields if necessary.
So for example you have to set `namespace` and `varnish_service` fields only once at the beginning of reconcile loop and all subsequent logging calls will have that field already set.
Also pass the context to functions that call Kubernetes API. In future we could easily add timeouts and deadlines and avoid hanging reconcile loops if that will start happening for some reason. Didn't set the timeout for now as I don't have a good value for that. After we measure the average reconciliation loop time we can set a reasonable value.
Set inject the version of the operator and kwatcher into the binary. That allowed to have the version in log entries and better identify the scope of the error.
Delete dead code.
Add logging function to put logger into context and retrieve from the context. Move logging fields into constants.

* Fix version injection.

* Remove reconciliation ID from log entries as it doesn't add enough value to justify its presence in each log entry.