Fix not working required fields validation. (#244)
* Fix not working required fields validation.

TL;DR The mutating webhook added the zero value (e.g. empty string) of the unspecified field making Kubernetes think that the field was specified.

The problem caused by mutating webhooks. Without it that .yaml manifest submitted by the user goes straight to Kubernetes and goes through the openapi schema validation. It does some simple validation defined in the CRD spec. So if the field is not specified, kubernetes does not see that field and says that it's not present and required.

With mutating webhooks and additional step is added **before** the schema validation. That step is the mutating webhook. Since the webhook is Go code, Go wold rules take place. The issue causing behavior is the way Go marshals JSON. Since the object being unmarshalled, mutated and the again marshalled, during marshalling the fields that does not include the `omitempty` tag will be present in the resulting JSON, even if the user did not specify it in his .yaml manifest. So in the end Kubernetes sees that field and passes the validation.

So the obvious fix would be to add `omitempty`. But kubebuilder uses it for generating the CRD. If `omitempty` is specified, that field is marked as optional. The opposite of what we want. The solution I found is to use the `// +kubebuilder:validation:Optional` marker on package level which essentially disables the behavior with `omitempty` tag and marks all the fields as optional in the generated CRD. Now, to mark the field as required you need to explicitly specify the marker `// +kubebuilder:validation:Required` for the field. Which is done here.

With those changes, the marshalled JSON will not include originally unspecified fields and all will work as expected. The downside of this solution is that you need to change a lot of field types to be a pointer. Only a nil values field can be omitted in the resulting JSON. Now you need to do additional checks for nil and default some values.

* Fix tests and varnishcontroller code.

* Make the service port also a pointer.