Create infrastructure for test and add clusterrole tests (#253)
* Fix not working required fields validation.

TL;DR The mutating webhook added the zero value (e.g. empty string) of the unspecified field making Kubernetes think that the field was specified.

The problem caused by mutating webhooks. Without it that .yaml manifest submitted by the user goes straight to Kubernetes and goes through the openapi schema validation. It does some simple validation defined in the CRD spec. So if the field is not specified, kubernetes does not see that field and says that it's not present and required.

With mutating webhooks and additional step is added **before** the schema validation. That step is the mutating webhook. Since the webhook is Go code, Go wold rules take place. The issue causing behavior is the way Go marshals JSON. Since the object being unmarshalled, mutated and the again marshalled, during marshalling the fields that does not include the `omitempty` tag will be present in the resulting JSON, even if the user did not specify it in his .yaml manifest. So in the end Kubernetes sees that field and passes the validation.

So the obvious fix would be to add `omitempty`. But kubebuilder uses it for generating the CRD. If `omitempty` is specified, that field is marked as optional. The opposite of what we want. The solution I found is to use the `// +kubebuilder:validation:Optional` marker on package level which essentially disables the behavior with `omitempty` tag and marks all the fields as optional in the generated CRD. Now, to mark the field as required you need to explicitly specify the marker `// +kubebuilder:validation:Required` for the field. Which is done here.

With those changes, the marshalled JSON will not include originally unspecified fields and all will work as expected. The downside of this solution is that you need to change a lot of field types to be a pointer. Only a nil values field can be omitted in the resulting JSON. Now you need to do additional checks for nil and default some values.

* Add more validation rules and cover them with tests.

* Remove validation for intstr types.
Add validation for service types.

* Add validation for ImagePullPolicy. Cover it with tests.

* Add validation for RestartPolicy. Cover it with tests.

* Cover it with tests a few more fields validation.

* Fix pointer issue. If you change a struct field passed to a function by reference the value of the parent struct will still have the old value.
Pass Varnish reconciler to the setup function for better testability.
Setup the test environment to start the test controlplane and the manager. The behavior is almost the same. So far the only observed difference is that the garbage collection doesnt work. Resources need to be cleaned up manually (a function for that is created).
Also since you don't have the kubernetes controllers (for statefulset, service, etc) and the helm installation part, some functionality like the admission webhooks, certs provisioning, etc. are not working.
For clusterrole and clusterrolebinding additional annotations were added. They include the information about the varnishcluster they were created for. It was done to fix an issue in watching functions. Since those 2 resource are clusterwide and not created in a namespace, when a change is observed, the controller runtime have no information about the namespace of the varnishcluster resource. By having those 2 annotations we can use the mapping function to map those resources to the corresponding varnishcluster.
Add tests for testing the clusterrole functionality. Its creation and update when a change is occurred (e.g. a user manually updated them).

* Update controller-gen tool to 2.0.4 and regenerate the CRD.

* Fix nil pointer dereference.

* Disable logging by default.